refactor(tests): strip redundant as-casts from shape test fixtures
TypeConstruction already carries its own type tag, typed RefDecls are in
the typeEnv, literals self-infer, and equality/??/(...)? all have
statically determinable return types. None of these sites need `as Type`.
DotAccessExpr annotations (p.x as Integer) are left intact — inference
doesn't yet handle field access on typed shapes.

+++ __tests__/types/shape_coercion.test.js
@@ -52,7 +52,7 @@ describe('shape coercion — typed assignment', () => {
       ::Point = (x Integer, y Integer)
       @go = {
         p Point = 1, 2, 3
-        -> result: 1 as Integer
+        -> result: 1
       }
     `)).toThrow();
   });
